Producer/requirements.txt
pymongo
kafka-python